Move preferences into lib directory
