+++ api/lib/spree/api.rb
@@ -8,5 +8,4 @@ module Spree
   end
 end
 
-require 'spree/api/api_dependencies'
 require 'spree/api/engine'
