+++ api/lib/spree/api/engine.rb
@@ -1,5 +1,8 @@
 require 'rails/engine'
 
+require_relative 'dependencies'
+require_relative 'configuration'
+
 module Spree
   module Api
     class Engine < Rails::Engine
@@ -7,8 +10,8 @@ module Spree
       engine_name 'spree_api'
 
       initializer 'spree.api.environment', before: :load_config_initializers do |_app|
-        Spree::Api::Config = Spree::ApiConfiguration.new
-        Spree::Api::Dependencies = Spree::ApiDependencies.new
+        Spree::Api::Config = Spree::Api::Configuration.new
+        Spree::Api::Dependencies = Spree::Api::ApiDependencies.new
       end
 
       initializer 'spree.api.checking_migrations' do
