+++ common_spree_dependencies.rb
@@ -64,6 +64,6 @@ group :test, :development do
 end
 
 group :development do
-  gem 'github_fast_changelog'
+  # gem 'github_fast_changelog'
   gem 'solargraph'
 end
