+++ core/lib/spree/core/engine.rb
@@ -1,3 +1,6 @@
+require_relative 'dependencies'
+require_relative 'configuration'
+
 module Spree
   module Core
     class Engine < ::Rails::Engine
@@ -19,7 +22,7 @@ module Spree
       end
 
       initializer 'spree.environment', before: :load_config_initializers do |app|
-        app.config.spree = Environment.new(SpreeCalculators.new, Spree::AppConfiguration.new, Spree::AppDependencies.new)
+        app.config.spree = Environment.new(SpreeCalculators.new, Spree::Core::Configuration.new, Spree::Core::Dependencies.new)
         Spree::Config = app.config.spree.preferences
         Spree::Dependencies = app.config.spree.dependencies
       end
