+++ core/lib/spree/core/preferences/preferable_class_methods.rb
