+++ core/lib/spree/core/preferences/scoped_store.rb
