+++ core/lib/spree/core/preferences/store.rb
