+++ core/spec/models/spree/app_dependencies_spec.rb
@@ -3,8 +3,8 @@ require 'spec_helper'
 class MyCustomCreateService
 end
 
-describe Spree::AppDependencies, type: :model do
-  let (:deps) { Spree::AppDependencies.new }
+describe Spree::Core::Dependencies, type: :model do
+  let (:deps) { described_class.new }
 
   it 'returns the default value' do
     expect(deps.cart_create_service).to eq('Spree::Cart::Create')
